This project is based on new image generation of Cat using GANs (Generative Adversarial Networks)
This project is based on new image generation of Cat using GANs (Generative Adversarial Networks) which is an architecture for training generative models such as deep convolutional neural networks for generating images. Developing a GAN for generating images requires a discriminator which is a convolutional neural network model for classifying whether a given image is real or fake (generated) and also a generator model that uses inverse convolutional layers to transform a random input noise to a full two-dimensional image of pixel values. Since, it is quite challenging to train and generate good quality color images of decent size using GANs, here, we have done a lot of hyper parameter tuning, changed the architecture of neural networks for both discriminator and generator many times and also tried 2 different datasets of cat images, one contains 2000 images and another contains 4000 images. Result with 4000 imgaes dataset was better though time-consuming. We resized all 4000 images to (64,64) before discriminator training. Model was run on GPU (on google colab) and we got decent generated cat images after 100 epochs.